requirements.txt
attrs==25.3.0
aws-cdk-lib==2.139.0
aws-cdk.asset-awscli-v1==2.2.251
aws-cdk.asset-kubectl-v20==2.1.4
aws-cdk.asset-node-proxy-agent-v6==2.1.0
cattrs==25.2.0
constructs==10.3.0
importlib_resources==6.5.2
jsii==1.114.1
publication==0.0.3
python-dateutil==2.9.0.post0
six==1.17.0
typeguard==2.13.3
typing_extensions==4.15.0
